refactor: util로 분리하지 않고 내부 함수에서 선언하는 방식으로 변경
